pwkit/msmt.py: tighten up handling of scalar Avals
Our various math functions often want to index into 'Aval.data', which isn't allowed
if that's a scalar. It's really annoying to have to implement two branches to everything
depending on whether the contents are a scalar or not. So instead store scalars as
shape (1,), and fake scalar-ness on interface functions like .shape or __len__().