Create the new "mathlib" modulo and wire it up.
To implement math on PKTable columns, I was going to need to duplicate a lot
of the kinds of "unary_math" infrastructure that I created for the Measurement
classes. I worked on this a bit and ended up havin to face a lot of corner
cases of, e.g., addition with various kinds of numerical-ish object types.

Prior work in this area has always made me think that it might be necessary to
split out the generic math library code (i.e., dispatching "log10" on
arbitrary objects), and getting the argument coercion to work right looked
error-prone enough that I finally did.

The new "mathlib" module mirrors all of Numpy's "ufunc" functions. Each of
these has infrastructure to delegate the implementation to a "math library"
class instance that implements some or all of the standard math functions.
Unsurprisingly there's baked-in support to delegate to Numpy.

This is still lightly tested but so far it seems to work and I think the
architecture is able to cover a wide range of uses. We'll see.