pwkit/mathlib.py: getting closer?
Made yet another pass at revising how the whole mathlib framework fits
together. I think this version is better and more manageable — it turns out
there are just lots and lots of corner cases to deal with. msmt and pktable
modules not yet update to deal with the changed interfaces.