Support objects with __array__ methods more transparently.
It felt like it was going to be a bit of a pain to implement a whole math
library just for PKTable ScalarColumns, which are really pretty dumb. Also,
this way we get free compatibility with objects like Pandas Series.

Since the extra logic comes in at the free-function wrapper level, we have to
make sure to invoke routines through those wrappers and not by direct calls on
MathFunctionLibrary instances, except in very controlled circumstances. So
get_library_for() has become quasi-private.