pwkit/mathlib.py: fix, uh, scalar math
The mathlib treats scalars like zero-dimensional arrays, i.e. arrays with
shape `()`. When doing math on such an "array" with "tidy"-based math
implementations, the tidying code reshapes it to shape `(1,)` so that the
underyling implementations can assume that the arrays are indexable. However,
it was still allocating the output array with the true array shape, `()`,
resulting in a dimension mismatch error. We fix this by semi-abusing some
variables that are reused in the relevant computations to make sure that the
output array also gets allocated with shape `(1,)`. Preexisting code then
reshapes *that* array to `()` before returning it to the caller.